Add optimization to breakends
